// client/src/components/stock/__generated__/StockLocations_query.graphql.ts
export type StockLocations_query$data = {
  readonly clogLocationSearch: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly stockCount: number;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "StockLocations_query";
};
export type StockLocations_query$key = {
  readonly " $data"?: StockLocations_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"StockLocations_query">;
};